src/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
SQLAlchemy
psycopg2-binary
openai==1.54.0
pandas==2.2.3
bcrypt<4.0.0
passlib[bcrypt]
python-jose
pyjwt==2.9.0
pydantic-settings==2.6.1
